latest code

+++ hni-admin-service/src/main/java/org/hni/admin/service/OrderController.java
@@ -15,7 +15,9 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import org.apache.commons.lang3.StringUtils;
+import org.hni.common.Constants;
 import org.hni.common.DateUtils;
+import org.hni.common.exception.HNIException;
 import org.hni.order.om.Order;
 import org.hni.order.om.OrderItem;
 import org.hni.order.service.OrderService;
@@ -25,6 +27,7 @@ import org.hni.provider.om.Menu;
 import org.hni.provider.om.MenuItem;
 import org.hni.provider.om.Provider;
 import org.hni.provider.om.ProviderLocation;
+import org.hni.provider.om.ProviderLocationHour;
 import org.hni.user.om.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -63,7 +66,10 @@ public class OrderController extends AbstractBaseController {
 		, response = Order.class
 		, responseContainer = "")
 	public String saveOrder(Order order) {
-		return serializeOrderToJson(orderService.save(order));
+		if (isPermitted(Constants.ORDER, Constants.CREATE, 0L)) {
+			return serializeOrderToJson(orderService.save(order));
+		}
+		throw new HNIException("You must have elevated permissions to do this.");
 	}
 
 	@DELETE
@@ -74,7 +80,10 @@ public class OrderController extends AbstractBaseController {
 		, response = Order.class
 		, responseContainer = "")
 	public String getDelete(@PathParam("id") Long id) {
-		return serializeOrderToJson(orderService.delete(new Order(id)));
+		if (isPermitted(Constants.ORDER, Constants.DELETE, 0L)) {
+			return serializeOrderToJson(orderService.delete(new Order(id)));
+		}
+		throw new HNIException("You must have elevated permissions to do this.");
 	}
 
 	@GET
@@ -115,14 +124,17 @@ public class OrderController extends AbstractBaseController {
 		, response = Order.class
 		, responseContainer = "")
 	public String resetOrder(@PathParam("id") Long id) {
-		if ( null != id ) {
-			Order order = orderService.get(id);
-			orderService.reset(order);
-			for(OrderPayment op : orderPaymentService.paymentsFor(order)) {
-				orderPaymentService.delete(op);
+		if (isPermitted(Constants.ORDER, Constants.CREATE, 0L)) {
+			if ( null != id ) {
+				Order order = orderService.get(id);
+				orderService.reset(order);
+				for(OrderPayment op : orderPaymentService.paymentsFor(order)) {
+					orderPaymentService.delete(op);
+				}
 			}
+			return String.format("{\"message\":\"reset order %d\"}", id);
 		}
-		return String.format("{\"message\":\"reset order %d\"}", id);
+		throw new HNIException("You must have elevated permissions to do this.");
 				
 	}
 
@@ -134,8 +146,11 @@ public class OrderController extends AbstractBaseController {
 		, response = Order.class
 		, responseContainer = "")
 	public void removeLock(@PathParam("id") Long id) {
-		logger.info("Unlocking Order "+id);
-		orderService.releaseLock(orderService.get(id));
+		if (isPermitted(Constants.ORDER, Constants.CREATE, 0L)) {
+			logger.info("Unlocking Order "+id);
+			orderService.releaseLock(orderService.get(id));
+		}
+		throw new HNIException("You must have elevated permissions to do this.");
 	}
 
 	private static final String ORDER_COUNT = "{\"order-count\":\"%d\"}";
@@ -176,6 +191,7 @@ public class OrderController extends AbstractBaseController {
 			String json = mapper.writeValueAsString(JsonView.with(order)
 					.onClass(User.class, Match.match().exclude("*").include("id", "firstName", "lastName"))
 					.onClass(ProviderLocation.class, Match.match().include("*").exclude("created", "createdById"))
+					.onClass(ProviderLocationHour.class, Match.match().exclude("*").include("dow", "openHour", "closeHour"))
 					.onClass(Provider.class, Match.match().include("*").exclude("created", "createdById"))
 					.onClass(OrderItem.class, Match.match().include("*").exclude("order"))
 					.onClass(Menu.class, Match.match().include("*").exclude("provider"))
